+++ R/read_write_option.R
@@ -24,87 +24,91 @@
 #' (read_write_option())
 #' # To add a new type:
 #' tail(read_write_option(data.frame(type = "png", read_fun = "png::readPNG",
-#'   read_header = NA, write_fun = "png::writePNG")))
+#'   read_header = NA, write_fun = "png::writePNG", comment = "PNG image")))
 read_write_option <- function(new_type) {
   opts <- getOption("read_write", default = tibble::tribble(
     ~type,     ~read_fun,              ~read_header,
-    ~write_fun,
+    ~write_fun,               ~comment,
     "csv",     "readr::read_csv",      "data.io::hread_text",
-    "readr::write_csv",
+    "readr::write_csv",       "comma separated values",
     "csv2",    "readr::read_csv2",     "data.io::hread_text",
-    NA,
+    NA,                       "semicolon separated values",
     "xlcsv",   "readr::read_csv",      "data.io::hread_text",
-    "readr::write_excel_csv",
+    "readr::write_excel_csv",  "write a CSV file more easily readable by Excel",
     "tsv",     "readr::read_tsv",      "data.io::hread_text",
-    "readr::write_tsv",
+    "readr::write_tsv",        "tab separated values",
     "fwf",     "readr::read_fwf",      "data.io::hread_text",
-    NA, # TODO: a writer here!
+    NA,                         "fixed width file", # TODO: a writer here!
     "log",     "readr::read_log",      NA,
-    NA, # TODO: a writer here!
+    NA,                          "standard log file", # TODO: a writer here!
     "rds",     "readr::read_rds",      NA,
-    "readr::write_rds",
+    "readr::write_rds",          "R data file (no compression by default)",
     "txt",     "readr::read_file",     NA,
-    "readr::write_file",
+    "readr::write_file",         "text file (as length 1 character vector)",
     "raw",     "readr::read_file_raw", NA,
-    NA, # TODO: a writer here!
+    NA,                           "binary file (read as raw vector)",
+                                                     # TODO: a writer here!
     "ssv",     "readr::read_table",    "data.io::hread_text",
-    NA,#Space separated values
+    NA,                           "space separated values (strict)",
     "ssv2",    "readr::read_table2",   "data.io::hread_text",
-    NA,
+    NA,                           "space separated values (relaxed)",
     "csv.gz",  "readr::read_csv",      "data.io::hread_text",
-    "readr::write_csv",
+    "readr::write_csv",            "gz compressed comma separated values",
     "csv2.gz", "readr::read_csv2",     "data.io::hread_text",
-    NA,
+    NA,                            "gz compressed semicolon separated values",
     "tsv.gz",  "readr::read_tsv",      "data.io::hread_text",
-    "readr::write_tsv",
+    "readr::write_tsv",             "gz compressed tab separated values",
     "txt.gz",  "readr::read_file",     NA,
-    "readr::write_file",
+    "readr::write_file",            "gz compressed text file",
     "csv.bz2", "readr::read_csv",      "data.io::hread_text",
-    "readr::write_csv",
+    "readr::write_csv",             "bz2 compressed comma separated values",
     "csv2.bz2","readr::read_csv2",     "data.io::hread_text",
-    NA,
+    NA,                             "bz2 compressed semicolon separated values",
     "tsv.bz2", "readr::read_tsv",      "data.io::hread_text",
-    "readr::write_tsv",
+    "readr::write_tsv",             "bz2 compressed tab separated values",
     "txt.bz2", "readr::read_file",     "data.io::hread_text",
-    "readr::write_file",
+    "readr::write_file",            "bz2 compressed text file",
     "csv.xz",  "readr::read_csv",      "data.io::hread_text",
-    "readr::write_csv",
+    "readr::write_csv",             "xz compressed comma separated values",
     "csv2.xz", "readr::read_csv2",     "data.io::hread_text",
-    NA,
+    NA,                             "xz compressed semicolon separated values",
     "tsv.xz",  "readr::read_tsv",      "data.io::hread_text",
-    "readr::write_tsv",
+    "readr::write_tsv",             "xz compressed tab separated values",
     "txt.xz",  "readr::read_file",     NA,
-    "readr::write_file",
+    "readr::write_file",            "xz compressed text file",
     # Buggy right now!! "csvy",    "csvy::read_csvy",    NA, "csvy::write_csvy",
+    # "comma separated value with YAML header",
     "xls",     "readxl::read_excel",   "data.io::hread_xls",
-    "WriteXLS::WriteXLS",
+    "WriteXLS::WriteXLS",            "Excel old .xls format",
     "xlsx",    "readxl::read_excel",   "data.io::hread_xlsx",
-    "openxlsx::write.xlsx",
+    "writexl::write_xlsx",    "Excel new .xlsx format", #"openxlsx::write.xlsx",
     "dta",     "haven::read_dta",      NA,
-    "haven::write_dta",
+    "haven::write_dta",              "Stata DTA format",
     # read_dta() = read_stata()
     "sas",     "haven::read_sas",      NA,
-    "haven::write_sas",
+    "haven::write_sas",               "SAS format",
     "sas7bdat","haven::read_sas",      NA,
-    "haven::write_sas",
+    "haven::write_sas",                "SAS format (sas7bdat)",
     "sav",     "haven::read_sav",      NA,
-    "haven::write_sav",
+    "haven::write_sav",                "SPSS .sav format",
+    "zsav",    "haven::read_sav",      NA,
+    "haven::write_sav",                "SPSS .zsav format",
     "por",     "haven::read_por",      NA,
-    NA,
+    NA,                                "SPSS .por format",
     # read_por()/read_sav() = read_spss()
     "xpt",     "haven::read_xpt",      NA,
-    "haven::write_xpt"   #,
+    "haven::write_xpt",                "SPSS transport format (FDA compliant)"#,
     #"feather", "feather::read_feather",NA,
-    #"feather::write_feather"
+    #"feather::write_feather",        "transportable feather format"
   ))
 
   if (!missing(new_type)) {
     # Check it is in a correct format
     if (!is.data.frame(new_type))
       stop("new_type must be a data.frame")
-    if (ncol(new_type) != 4)
-      stop("new_type must contain 4 columns",
-        " (type, read_fun, read_header & write_fun")
+    if (ncol(new_type) != 5)
+      stop("new_type must contain 5 columns",
+        " (type, read_fun, read_header, write_fun & comment")
     names(new_type) <- names(opts)
     opts <- rbind(opts, new_type)
   }