+++ NEWS.md
@@ -1,8 +1,18 @@
 # data.io News
 
+## Changes in data.io 1.1.0
+
+- A basic version of write() is now available.
+
+- data_types() function added to easily get information about data types that
+  can be read() or write()
+
+- Description added into "read_write" options.
+
+
 ## Changes in data.io 1.0.1
 
-- bug corrected: forgot to change 'data' -> 'data.io' in 'read_write' options.
+- Bug corrected: forgot to change 'data' -> 'data.io' in 'read_write' options.
 
 
 ## Changes in data.io 1.0.0